Create LinkedListImplementation.cpp
Create s a linked list of US cities. Each entry in the list is to have a city name, city population, maiden household income and a pointer to the next city record. Also, the head pointer points at the beginning of the list and the tail pointer is pointing at the last node of the list.